refactor: 선택 가능 숫자, 선택한 숫자 일급 컬렉션 분리
