+++ src/test/java/lotto/Fixtures.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 public class Fixtures {
 
-    public static Numbers createNumbers(List<Integer> numbers) {
-        return new Numbers(numbers.stream().map(Number::new).collect(Collectors.toList()));
+    public static PickedNumbers createNumbers(List<Integer> numbers) {
+        return new PickedNumbers(numbers.stream().map(Number::new).collect(Collectors.toList()));
     }
 }
