+++ src/test/java/lotto/PickedNumbersTest.java
@@ -8,7 +8,7 @@ import java.util.Collections;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class NumbersTest {
+public class PickedNumbersTest {
 
     @Test
     @DisplayName("6개의 로또 번호로 로또 번호 일급 컬렉션을 생성할 수 있다.")
@@ -19,7 +19,7 @@ public class NumbersTest {
     @Test
     @DisplayName("로또 번호 갯수는 6보다 작거나 클 수 없다.")
     void create_invalidSize() {
-        assertThatThrownBy(() -> new Numbers(Collections.emptyList()))
+        assertThatThrownBy(() -> new PickedNumbers(Collections.emptyList()))
             .isInstanceOf(IllegalArgumentException.class);
     }
 }
